requirements.txt
# Dependencies
Django>=4.2.8
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-redis>=5.4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djoser>=2.2.0
Pillow>=10.0.0
django-filter>=23.5
django-celery-beat>=2.5.0
celery>=5.3.6
redis>=5.0.1
whitenoise>=6.6.0
urllib3>=2.0.7
python-dateutil>=2.8.2

# Code quality tools
flake8>=6.1.0
black>=23.11.0